copy files if required for scene object loading
- extract prepareFilepathForModel
