+++ src/js/shot-generator/Editor.js
@@ -630,6 +630,8 @@ const SceneManager = connect(
 
                 loaded: props.loaded ? props.loaded : false,
 
+                storyboarderFilePath: meta.storyboarderFilePath,
+
                 ...props
               }
             ]
